add main menu
