+++ app/src/at-longphan/java/com/asiantech/intern20summer1/week6/activity/SecondPageActivity.kt
@@ -18,7 +18,7 @@ class SecondPageActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_second_page)
+        setContentView(R.layout.activity_second_page_w6)
         configStatusBar()
         initViews()
         setupViewPager()
