+++ app/src/at-longphan/java/com/asiantech/intern20summer1/week6/fragment/ThirdStepFragment.kt
@@ -40,7 +40,7 @@ class ThirdStepFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val view: View = inflater.inflate(R.layout.fragment_step_3, container, false)
+        val view: View = inflater.inflate(R.layout.fragment_step_3_w6, container, false)
         val tvStep = view.findViewById(R.id.tvStepDescription) as? TextView
         tvStep?.text = step
