+++ app/src/at-longphan/res/layout/fragment_random_image_w6.xml
